# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Salon Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Get revenue report
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
components:
  schemas:
...
    RevenueDataPoint:
      type: object
      properties:
        label:
          type: string
        revenue:
          type: number
        bills:
          type: number
        customers:
          type: number
      required: [label, revenue]
    RevenueReport:
      type: object
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        totalBills:
          type: number
        data:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
      required: [period, totalRevenue, totalBills, data]